The rest of the commit, because I thought of sharing this and I started working on some other repo :|
So, no history on this one.

If you are interested in a good comment here then, take a seat kids for I am about
to tell you the story of how I met your mother.

... So, the project is pretty self explaining, the FunkyAnimationGradient contains
most of the logic. I prefer working with observed objects as opposed to state, so pulled
that out and end up with a pretty view model that triggers after a few seconds to give
you an animation of colours flickering through the screen.

The gradient starts in a diagonal, we quickly start shifting things around to provide
something more interesting.